Launcher3: Introduce overview scrolling scale animation
This vivid animation enables users to perceive more clearly
about the switching of the task cards.

Note that live tile feature is disabled because it is not
compatible with this animation. Also live tile is causing
notable performance regression, which is not worth relative
to its function.

Change-Id: Ie1350d6bda2d311f70f9be7cff47b0abe7e00cad
Signed-off-by: LibXZR <i@xzr.moe>

Fix discontinuous animation when dismissing taskcards on
 the right side

Cards should never be scale down in this animation.

Change-Id: I759492390eb8f1a2b9023c2d264f0e7191b17108
Signed-off-by: LibXZR <i@xzr.moe>

Recalculate scroll scale on layout

Edge case exist that when there's only one task,
onScrollChanged() won't be called so the last scale
from the recycled view is showed, leading to wrong
display size.

Change-Id: If8a189ba6b854ed486283d43b8a0ac3e4bb4e9f2
Signed-off-by: LibXZR <i@xzr.moe>

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Anush02198 <anush.4376@gmail.com>